Update README.md
We pass rows as STDIN however our description mentions the input.txt file a lot which can be very confusing and hence I see many submissions reading from file and also, writing to a file. This leads to us manually making the change in the code which in turn means that we can't judge the submission on the way they handle input.